+++ frontend/helpers/string_helpers.js
@@ -1,0 +1,21 @@
+export const processMessages = (array, statusCode = 200) => {
+  let type = 'error';
+
+  if (statusCode >= 200 && statusCode < 300) {
+    type = 'success';
+  }
+
+  return array.map(message => ({ type, message }));
+};
+
+export const hashString = string => {
+  let hash = 0;
+  let i = 0;
+  let len = string.length;
+
+  while (i < len) {
+    hash = ((hash << 5) - hash + string.charCodeAt(i++)) << 0;
+  }
+
+  return hash;
+};
